Fix MoboxExplorer.desktop icon, add Shutdown.desktop
The icon for MoboxExplorer.desktop was set to "wine". On a fresh termux native installation, there is no icon for wine. Therefore, I had set it to
a gaming controller instead.

I have also added a new desktop file called "Shutdown.desktop" which aims to make it simpler to shutdown Termux X11 within it itself for convenience